added cosine EMD distance
